[34] #fixes 34 Adds Typescript
