+++ api-client/README.md
@@ -11,3 +11,9 @@ yarn install
 ```
 yarn start
 ```
+
+# Routes
+
+- [http://localhost:3000/](http://localhost:3000/)
+- [http://localhost:3000/api/v1/quiz/](http://localhost:3000/api/v1/quiz/)
+- [http://localhost:3000/api/v1/quiz/:slug](http://localhost:3000/api/v1/quiz/javascript-quiz)
