+++ api-client/api-models.ts
@@ -1,0 +1,5 @@
+export interface QuizDTO {
+    id: string
+    title: string
+    slug: string
+}
