+++ api-client/mockData/index.ts
@@ -1,0 +1,7 @@
+import quiz from "./quiz"
+
+const data = {
+    quiz: quiz
+}
+
+export default data
