+++ api-client/mockData/quiz.ts
@@ -1,0 +1,22 @@
+import { QuizDTO } from "../api-models"
+
+const quiz: QuizDTO[] = [
+  {
+    id: "1",
+    title: "HTML Quiz",
+    slug: 'html-quiz',
+  },
+  {
+    id: "2",
+    title: "JavaScript Quiz",
+    slug: 'javascript-quiz',
+  },
+  {
+    id: "3",
+    title: "CSS Quiz",
+    slug: 'css-quiz',
+  },
+
+]
+
+export default quiz
