+++ api-client/package.json
@@ -7,9 +7,13 @@
   "license": "MIT",
   "private": false,
   "scripts": {
-    "start": "node src/server.js"
+    "start": "npx ts-node src/server.ts"
   },
   "devDependencies": {
-    "json-server": "^0.17.0"
+    "json-server": "^0.17.0",
+    "@types/json-server": "^0.14.4"
+  },
+  "dependencies": {
+    "typescript": "^4.8.3"
   }
 }
